change int field types tro bigint, for identity columns

+++ publishable/database/migrations/2017_11_05_060225_create_admin_settings_table.php
@@ -25,7 +25,7 @@ class CreateAdminSettingsTable extends Migration
 	              ->references('id')
 	              ->on('admin_setting_groups');
 
-	        $table->integer('crud_field_type_id')
+	        $table->bigInteger('crud_field_type_id')
 	              ->unsigned()
 	              ->nullable();
